chore: more refactoring/renaming
