+++ packages/ponder/src/subgraph-compat/esbuildPlugin.ts
