+++ packages/ponder/src/subgraph-compat/readMappings.ts
