+++ packages/ponder/src/subgraph-compat/readSubgraphSchema.ts
